\name{eDNAShinyApp-package}
\alias{eDNAShinyApp-package}
\alias{eDNAShinyApp}
\docType{package}
\title{\packageTitle{eDNAShinyApp}}
\description{\packageDescription{eDNAShinyApp}}
\section{Package Content}{\packageIndices{eDNAShinyApp}}
\author{\packageAuthor{eDNAShinyApp}}
\section{Maintainer}{\packageMaintainer{eDNAShinyApp}}
\keyword{package}
